requirements.txt
fastapi[standard]
py-mon
requests
python-dotenv
unidecode
pydantic
inflection
pymongo
invoke
vnstock3